+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [7.14.0] 2025-12-10
+
+### Added
+
+- [TD-7711]:
+  - Phrase prefix search to highlight stricter text matches
+  - Remove ingest search
+
 ## [7.13.0] 2025-11-18
 
 ### Added